+++ src/main.cpp
@@ -101,16 +101,17 @@ int internal::main(int argc, char** argv) {
   std::variant<avif::img::Image<8>, avif::img::Image<16>> loadedImage = PNGReader::create(config.input).read();
 
   aom_image_t img;
-  avif::img::ColorProfile colorProfile;
+  avif::img::ColorProfile colorProfileFromImage = std::monostate{};
   if(std::holds_alternative<avif::img::Image<8>>(loadedImage)) {
     auto src = std::get<avif::img::Image<8>>(loadedImage);
     convert(config, src, img);
-    colorProfile = src.colorProfile();
+    colorProfileFromImage = src.colorProfile();
   } else {
     auto src = std::get<avif::img::Image<16>>(loadedImage);
     convert(config, src, img);
-    colorProfile = src.colorProfile();
+    colorProfileFromImage = src.colorProfile();
   }
+  config.validate();
 
   uint32_t const width = aom_img_plane_width(&img, AOM_PLANE_Y);
   uint32_t const height = aom_img_plane_height(&img, AOM_PLANE_Y);
@@ -196,7 +197,7 @@ int internal::main(int argc, char** argv) {
     if (!seq.has_value()) {
       throw std::logic_error("No sequence header OBU.");
     }
-    builder.setPrimaryFrame(AVIFBuilder::Frame(colorProfile, seq.value(), std::move(configOBUs), std::move(mdat)));
+    builder.setPrimaryFrame(AVIFBuilder::Frame(colorProfileFromImage, seq.value(), std::move(configOBUs), std::move(mdat)));
     if(config.alphaInput.has_value()) {
       log.info("Attaching {} as Alpha plane.", config.alphaInput.value());
       builder.setAlphaFrame(AVIFBuilder::Frame::load(log, config.alphaInput.value()));