+++ CHANGELOG.md
@@ -25,8 +25,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - libaom is upgraded to v3.2.0
 - libvmaf is upgrade to [441ab02a6b9df77a716f9bc1772340acd36e201b](https://github.com/Netflix/vmaf/tree/441ab02a6b9df77a716f9bc1772340acd36e201b).
 - [Remove URL from header to make image sizes smaller](https://github.com/link-u/cavif/pull/56).
+- Default values of `--color-primaries`, `--transfer-characteristics` and `--matrix-coefficients` are now "empty".
+  - If those flags are not set and input PNG file does not contain a `sRGB` chunk, these settings are used:
+    - `--color-primaries 1` (sRGB/sYCC)
+    - `--transfer-characteristics 13` (sRGB/sYCC)
+    - `--matrix-coefficients`
 
 ### Fixed
+
 - Now we can build cavif for mac OS and debian package.
+- Default
 
 ### Removed